Renamed Calendar to CalendarWeekly
